"Added custom hooks for handling oauth"
